requirements.txt
psutil>=5.9.0
requests>=2.31.0
click>=8.1.0
tqdm>=4.65.0
python-dateutil>=2.8.2 